test: share strict test utilities
