+++ test/lib/zwave/AssociationService.test.ts
@@ -11,6 +11,7 @@ import type {
 	AssociationNodeState,
 	AssociationZWaveNodeHandle,
 } from '../../../api/lib/zwave/ports.ts'
+import { createDeferred } from '../testUtils.ts'
 
 function makeZWaveNode(
 	overrides: Partial<{
@@ -115,20 +116,6 @@ function createService(
 	const log = createLogPort()
 	const service = new AssociationService(driver, nodeStore, log)
 	return { service, driver, nodeStore, log }
-}
-
-function createDeferred<T = void>(): {
-	promise: Promise<T>
-	resolve: (value: T | PromiseLike<T>) => void
-	reject: (reason?: unknown) => void
-} {
-	let resolve!: (value: T | PromiseLike<T>) => void
-	let reject!: (reason?: unknown) => void
-	const promise = new Promise<T>((res, rej) => {
-		resolve = res
-		reject = rej
-	})
-	return { promise, resolve, reject }
 }
 
 function createMutableDriverPort(
